product-issue-miner/mvp/requirements.txt
# MVP Dependencies
httpx>=0.27.0          # Async HTTP client for Zendesk
anthropic>=0.40.0      # Claude AI
python-dotenv>=1.0.0   # Environment variables
rich>=13.0.0           # Nice CLI output
fastapi>=0.115.0       # Web API
uvicorn>=0.32.0        # ASGI server
